requirements.txt
annotated-types==0.7.0
cohere==5.16.2
email-validator==2.3.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
langchain==0.3.27
langchain-cohere==0.4.5
langchain-community==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-openai==0.3.30
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
numpy==2.3.2
openai==1.99.9
pandas==2.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
regex==2025.7.34
streamlit==1.49.1
tokenizers==0.21.4